Move schema related logic to separate file
